VMM: VC++ 19.2 update 11 build adjustments (lossy floating point conversions). bugref:8489
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@93308 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/VMM/VMMR3/PGMSavedState.cpp
@@ -1906,7 +1906,7 @@ static DECLCALLBACK(int)  pgmR3LiveVote(PVM pVM, PSSMHANDLE pSSM, uint32_t uPass
 
     /* estimate the speed */
     uint64_t cNsElapsed = RTTimeNanoTS() - pVM->pgm.s.LiveSave.uSaveStartNS;
-    uint32_t cPagesPerSecond = (uint32_t)(   pVM->pgm.s.LiveSave.cSavedPages
+    uint32_t cPagesPerSecond = (uint32_t)(  (long double)pVM->pgm.s.LiveSave.cSavedPages
                                           / ((long double)cNsElapsed / 1000000000.0) );
     pVM->pgm.s.LiveSave.cPagesPerSecond = cPagesPerSecond;
